Refactor StackResolver class into separate files
This commit also adds the DbgHelp / SymSrv modules - not yet wired up to
rest of the code flow.
