+++ Engine/packages.config
@@ -4,7 +4,7 @@
   <package id="Microsoft.CodeAnalysis.VersionCheckAnalyzer" version="3.0.0" targetFramework="net472" developmentDependency="true" />
   <package id="Microsoft.CodeQuality.Analyzers" version="3.0.0" targetFramework="net472" developmentDependency="true" />
   <package id="Microsoft.Diagnostics.NETCore.Client" version="0.2.137102" targetFramework="net472" />
-  <package id="Microsoft.Diagnostics.Runtime" version="2.0.137201" targetFramework="net472" />
+  <package id="Microsoft.Diagnostics.Runtime" version="2.0.161401" targetFramework="net472" />
   <package id="Microsoft.NetCore.Analyzers" version="3.0.0" targetFramework="net472" developmentDependency="true" />
   <package id="Microsoft.NetFramework.Analyzers" version="3.0.0" targetFramework="net472" developmentDependency="true" />
   <package id="Microsoft.SqlServer.XEvent.XELite" version="2019.11.20.2" targetFramework="net472" />
